+++ src/main/java/golden/botc_mc/botc_mc/TitleUtil.java
@@ -1,11 +1,17 @@
 package golden.botc_mc.botc_mc;
 
+import net.minecraft.component.DataComponentTypes;
+import net.minecraft.component.type.DeathProtectionComponent;
+import net.minecraft.item.ItemStack;
 import net.minecraft.network.packet.s2c.play.OverlayMessageS2CPacket;
 import net.minecraft.network.packet.s2c.play.SubtitleS2CPacket;
 import net.minecraft.network.packet.s2c.play.TitleFadeS2CPacket;
 import net.minecraft.network.packet.s2c.play.TitleS2CPacket;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.text.Text;
+import net.minecraft.util.Hand;
+
+import java.util.List;
 
 /**
  * Utility class for showing titles, subtitles, and action bar messages to players.
@@ -101,6 +107,28 @@ public class TitleUtil {
     private static void setTimings(ServerPlayerEntity player, int fadeInTicks, int stayTicks, int fadeOutTicks) {
         TitleFadeS2CPacket titleFadeS2CPacket = new TitleFadeS2CPacket(fadeInTicks, stayTicks, fadeOutTicks);
         player.networkHandler.sendPacket(titleFadeS2CPacket);
+    }
+
+    /**
+     * Show a totem pop effect to a player with a custom item.
+     * @param player The player to show the totem effect to.
+     * @param item The item to show as the effect.
+     */
+    public static void showTotemEffect(ServerPlayerEntity player, ItemStack item) {
+        DeathProtectionComponent deathProtectionComponent = new DeathProtectionComponent(List.of());
+        item.set(DataComponentTypes.DEATH_PROTECTION, deathProtectionComponent);
+
+        // Store the item in the player's offhand to restore later
+        ItemStack previousOffhand = player.getOffHandStack();
+        // Set the offhand item to the pop item stack
+        player.setStackInHand(Hand.OFF_HAND, item);
+        // Send the item update to the client to ensure it sees the new offhand item before the pop effect
+        player.currentScreenHandler.sendContentUpdates();
+        // Trigger the totem pop effect
+        player.getWorld().sendEntityStatus(player, (byte) 35);
+        // Restore the player's original offhand item after the pop effect
+        player.setStackInHand(Hand.OFF_HAND, previousOffhand);
+        player.currentScreenHandler.sendContentUpdates();
     }
 
 }
